score-function now working properly

+++ include/score.h
@@ -7,6 +7,6 @@
 #include <iostream>
 #include "Sequence.h"
 
-int score(std::vector<std::tuple<unsigned long long,int, int> >& matches, Sequence& seq1, Sequence& seq2, std::string pattern, int& dc, int& threshold);
+double score(std::vector<std::tuple<unsigned long long,int, int> >& matches, Sequence& seq1, Sequence& seq2, std::string pattern, int& dc, int& threshold);
 
 #endif // SCORE_H
